+++ app/src/main/java/com/anytypeio/anytype/di/main/WorkspaceModule.kt
@@ -5,6 +5,7 @@ import com.anytypeio.anytype.data.auth.status.SyncAndP2PStatusEventsStore
 import com.anytypeio.anytype.domain.base.AppCoroutineDispatchers
 import com.anytypeio.anytype.domain.block.repo.BlockRepository
 import com.anytypeio.anytype.domain.debugging.Logger
+import com.anytypeio.anytype.domain.notifications.NotificationBuilder
 import com.anytypeio.anytype.domain.workspace.SpaceManager
 import com.anytypeio.anytype.domain.workspace.SyncAndP2PStatusChannel
 import dagger.Module
@@ -20,11 +21,13 @@ object WorkspaceModule {
     fun spaces(
         repo: BlockRepository,
         dispatchers: AppCoroutineDispatchers,
-        logger: Logger
+        logger: Logger,
+        notificationBuilder: NotificationBuilder
     ): SpaceManager = SpaceManager.Impl(
         dispatchers = dispatchers,
         repo = repo,
-        logger = logger
+        logger = logger,
+        notificationBuilder = notificationBuilder
     )
 
     @JvmStatic
